🧩 :: RecruitmentDetailViewController 수정
